# openapi.yaml
openapi: 3.0.3
info:
  title: NEXOHouse
  description: NEXOHouse documentation
  version: 1.0.0
servers:
  - url: https://api.nexo-house.com/api/v1
  - url: http://localhost:9000/api/v1
...
paths:
  /auth:
    post:
      tags:
        - Auth
      summary: Auth user
      description: Auth user.
      operationId: authUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: pekrloloe@gmail.com
                password:
                  type: string
                  example: "12345678"
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
    get:
      tags:
        - Auth
      summary: Geat All Users
      security:
        - bearerAuth: []
      description: Geat All Users
      operationId: getAllUsers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usersget"
...
components:
  schemas:
    User:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Test
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: password
        profile:
          type: string
          enum:
            - DataEntry
            - Administrator
          default: DataEntry
        church:
          type: string
          example: 63ccad4fac7871cc2f7b84cb
        rol:
          type: string
          enum:
            - SuperAdministrator
            - Administrator
            - Responsible
          default: Responsible
        address:
          type: string
          example: "Calle # 12"
        phone:
          type: string
          example: "+5258493624"
...
    Usersget:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          type: object
          properties:
            pagination:
              type: object
              properties:
                prev:
                  example: 1 || false
                page:
                  example: 1
                next:
                  example: 2 || false
                per_page:
                  example: 10
                total_pages:
                  example: 3
                per_page:
                  total_docs: 30
            info:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: Andres
                  email:
                    type: string
                    example: andres@nexolife.com
                  profile:
                    type: string
                    enum:
                      - DataEntry
                      - Administrator
                    default: DataEntry
                  church:
                    type: string
                    example: 63ccad4fac7871cc2f7b84cb
                  rol:
                    type: string
                    enum:
                      - SuperAdministrator
                      - Administrator
                      - Responsible
                    default: Responsible
                  status:
                    type: boolean
                    example: true
                  address:
                    type: string
                    example: "Calle # 12"
                  phone:
                    type: string
                    example: "+5258493624"
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
...